+++ lib/graphql/transaction.ts
@@ -4,8 +4,34 @@ import {
   Transaction as TransactionEntry
 } from "./schema";
 import { IterableModel } from "./iterableModel";
-import { FetchOptions } from "./types";
+import {
+  FetchOptions,
+  SubscriptionType,
+  Subscription
+} from "./types";
 import { ResultPage } from "./resultPage";
+
+const TRANSACTION_FIELDS = `
+  id
+  amount
+  name
+  iban
+  type
+  bookingDate
+  valutaDate
+  originalAmount
+  foreignCurrency
+  e2eId
+  mandateNumber
+  paymentMethod
+  category
+  userSelectedBookingDate
+  purpose
+  documentNumber
+  documentPreviewUrl
+  documentDownloadUrl
+  documentType
+`;
 
 const FETCH_TRANSACTIONS = `query fetchTransactions ($first: Int, $last: Int, $after: String, $before: String) {
   viewer {
@@ -13,25 +39,7 @@ const FETCH_TRANSACTIONS = `query fetchTransactions ($first: Int, $last: Int, $a
       transactions(first: $first, last: $last, after: $after, before: $before) {
         edges {
           node {
-              id
-              amount
-              name
-              iban
-              type
-              bookingDate
-              valutaDate
-              originalAmount
-              foreignCurrency
-              e2eId
-              mandateNumber
-              paymentMethod
-              category
-              userSelectedBookingDate
-              purpose
-              documentNumber
-              documentPreviewUrl
-              documentDownloadUrl
-              documentType
+            ${TRANSACTION_FIELDS}
           }
         }
         pageInfo {
@@ -45,6 +53,12 @@ const FETCH_TRANSACTIONS = `query fetchTransactions ($first: Int, $last: Int, $a
   }
 }`;
 
+export const NEW_TRANSACTION_SUBSCRIPTION = `subscription {
+  newTransaction {
+    ${TRANSACTION_FIELDS}
+  }
+}`;
+
 export class Transaction extends IterableModel<TransactionEntry> {
   /**
    * Fetches first 50 transactions which match the query
@@ -55,9 +69,26 @@ export class Transaction extends IterableModel<TransactionEntry> {
   async fetch(args?: FetchOptions): Promise<ResultPage<TransactionEntry>> {
     const result: Query = await this.client.rawQuery(FETCH_TRANSACTIONS, args);
 
-    const transactions = (result?.viewer?.mainAccount?.transactions?.edges ?? []).map((edge: TransactionsConnectionEdge) => edge.node);
+    const transactions = (
+      result?.viewer?.mainAccount?.transactions?.edges ?? []
+    ).map((edge: TransactionsConnectionEdge) => edge.node);
 
-    const pageInfo = result?.viewer?.mainAccount?.transactions?.pageInfo ?? { hasNextPage: false, hasPreviousPage: false};
+    const pageInfo = result?.viewer?.mainAccount?.transactions?.pageInfo ?? {
+      hasNextPage: false,
+      hasPreviousPage: false
+    };
     return new ResultPage(this, transactions, pageInfo);
+  }
+
+  subscribe(
+    onNext: (event: TransactionEntry) => any,
+    onError?: (error: Error) => any
+  ): Subscription {
+    return this.client.subscribe({
+      query: NEW_TRANSACTION_SUBSCRIPTION,
+      type: SubscriptionType.newTransaction,
+      onNext,
+      onError
+    });
   }
 }